+++ delete.php
@@ -1,14 +1,12 @@
 <?php
 require "config/database.php";
 
-if(isset($_GET['id']))
-{
-    $id = $_GET["id"];
+if (isset($_GET['id']) && is_numeric($_GET['id'])) {
+    $id = (int) $_GET['id'];
     $sql = "DELETE FROM article WHERE id = ?";
     $req = $db->prepare($sql);
     $req->execute([$id]);
-    header("location:header.php");
 }
-else{
-    echo "aucun id";
-}
+
+header("location:article.php");
+exit;
